refactored filenames and edited spring config file
