+++ source/projects/demos/LoggingModifiers/src/edu/emory/cci/sample/modifiers/LoggingQM.java
@@ -1,7 +1,5 @@
 package edu.emory.cci.sample.modifiers;
 
-import java.io.ByteArrayInputStream;
-import java.io.IOException;
 import java.util.Map;
 
 import org.apache.commons.logging.Log;
@@ -9,16 +7,12 @@ import org.apache.commons.logging.LogFactory;
 
 import com.google.gson.JsonObject;
 
-import edu.emory.cci.bindaas.framework.api.IQueryResultModifier;
+import edu.emory.cci.bindaas.framework.api.IQueryModifier;
 import edu.emory.cci.bindaas.framework.model.ModifierException;
-import edu.emory.cci.bindaas.framework.model.QueryResult;
 import edu.emory.cci.bindaas.framework.model.RequestContext;
 import edu.emory.cci.bindaas.framework.provider.exception.AbstractHttpCodeException;
-import edu.emory.cci.bindaas.framework.provider.exception.QueryExecutionFailedException;
-import edu.emory.cci.bindaas.framework.util.IOUtils;
-import edu.emory.cci.bindaas.framework.util.StandardMimeType;
 
-public class LoggingQM implements IQueryResultModifier{
+public class LoggingQM implements IQueryModifier{
 
 	private Log log = LogFactory.getLog(getClass());
 	@Override
@@ -26,40 +20,33 @@ public class LoggingQM implements IQueryResultModifier{
 
 		return new JsonObject();
 	}
-
 	@Override
 	public void validate() throws ModifierException {
-		// not implemented
-		
+		// not implemented	
 	}
 
 	@Override
 	public String getDescriptiveName() {
 
-		return "Plugin for logging query results";
+		return "Plugin for logging query";
 	}
 
 	@Override
-	public QueryResult modifyQueryResult(QueryResult queryResult,
-			JsonObject dataSource, RequestContext requestContext,
-			JsonObject modifierProperties, Map<String, String> queryParams)
+	public String modifyQuery(String query, JsonObject dataSource,
+			RequestContext requestContext, JsonObject modifierProperties)
 			throws AbstractHttpCodeException {
-		String data;
-		try {
-			
-			// intercepting the response and logging it
-			data = IOUtils.toString(queryResult.getData());
-			log.info("Request received from [" + requestContext.getUser() + "]");
-			log.info("Response [" + data + "]");
-			// overriding the response by a custom message
-			queryResult.setMimeType(StandardMimeType.TEXT.toString());
-			queryResult.setData(new ByteArrayInputStream(new String("Response from the QueryHandler was consumed by LoggingQRM").getBytes()));
-			return queryResult;
-		} catch (IOException e) {
-			log.error("Execution of LoggingQRM failed");
-			throw new QueryExecutionFailedException(getClass().getName(), 1);
-		}
-
+		log.info("Query  intercepted [" + query + "]");
+		return query;
 	}
 
+	@Override
+	public Map<String, String> modiftQueryParameters(
+			Map<String, String> queryParams, JsonObject dataSource,
+			RequestContext requestContext, JsonObject modifierProperties)
+			throws AbstractHttpCodeException {
+		
+		return queryParams;
+	}
+
+	
 }
